+++ src/pages/app/components/LayoutAppPaperbase.jsx
@@ -34,7 +34,7 @@ function LayoutAppPaperbase({ children }) {
     const [mobileOpen, setMobileOpen] = React.useState(false)
     const isSmUp = useMediaQuery(theme.breakpoints.up('sm'))
 
-    const user = auth.currentUser || {}
+    // const user = auth?.currentUser || {}
 
     // const localRole = localStorage.getItem('role')
     // console.log(JSON.parse(localRole))
