+++ src/pages/app/mensajes/index.page.jsx
@@ -4,7 +4,7 @@ export { Page }
 import React from 'react'
 import es from 'date-fns/locale/es'
 import { auth } from '#@/firebase/firebaseClient'
-import { App as SendBirdApp } from '@sendbird/uikit-react'
+// import { App as SendBirdApp } from '@sendbird/uikit-react'
 // react-bootrstrap
 import Row from 'react-bootstrap/Row'
 import Col from 'react-bootstrap/Col'
@@ -16,7 +16,7 @@ import Container from 'react-bootstrap/Container'
 // import TextareaAutosize from '@mui/material/TextareaAutosize'
 
 const Page = () => {
-    let userId = auth.currentUser.uid || undefined
+    let userId = auth?.currentUser?.uid || undefined
     let appId = import.meta.env.VITE_APP_SENDBIRD_APPID
     // let appId = process.env.VITE_APP_SENDBIRD_APPID
 
@@ -31,11 +31,11 @@ const Page = () => {
                         style={{ height: '50vh' }}
                     >
                         <h2 className="headline-xl">Mensajes</h2>
-                        <SendBirdApp
+                        {/* <SendBirdApp
                             appId={appId} // Sendbird application ID.
                             userId={userId} // user ID.
                             dateLocale={es}
-                        />
+                        /> */}
                         {/* <FormControl
                             fullWidth
                             style={{
