+++ src/pages/app/renderer/Link.tsx
@@ -1,15 +1,15 @@
-import PropTypes from 'prop-types'
+// import PropTypes from 'prop-types'
 import { usePageContext } from './usePageContext'
 
 export { Link }
 
-Link.propTypes = {
-    className: PropTypes.string,
-    href: PropTypes.string.isRequired,
-}
-function Link(props) {
+function Link(props: {
+    href?: string
+    className?: string
+    children: React.ReactNode
+}) {
     const pageContext = usePageContext()
-    console.log('function Link', pageContext.urlPathname, props.href) // .split('/')[3]
+    // console.log('function Link', pageContext.urlPathname, props.href) // .split('/')[3]
     const className = [
         props.className,
         pageContext.urlPathname === props.href && 'is-active',
