+++ src/pages/app/renderer/PageShell.tsx
@@ -1,8 +1,6 @@
 export { PageShell }
 
 import React from 'react'
-// import PropTypes from 'prop-types'
-// import { childrenPropType } from './PropTypeValues'
 import type { PageContext } from './types'
 import { LayoutAppPaperbase } from '#@/pages/app/components/LayoutAppPaperbase'
 import { PageContextProvider } from './usePageContext'
@@ -12,10 +10,6 @@ import { Providers } from './Providers'
 import '#@/pages/app/renderer/Private-App.scss'
 import '#R/index.scss'
 
-// PageShell.propTypes = {
-//     pageContext: PropTypes.any,
-//     children: childrenPropType,
-// }
 
 function PageShell({
     children,
@@ -24,7 +18,6 @@ function PageShell({
     children: React.ReactNode
     pageContext: PageContext
 }) {
-    // const Layout = LayoutPaperbase || pageContext.exports.Layout
 
     return (
         <React.StrictMode>
