--- src/pages/app/renderer/PropTypeValues.js
@@ -1,6 +1,0 @@
-import PropTypes from 'prop-types'
-
-export let childrenPropType = PropTypes.oneOfType([
-    PropTypes.arrayOf(PropTypes.node),
-    PropTypes.node,
-]).isRequired
