+++ src/pages/app/renderer/usePageContext.tsx
@@ -2,8 +2,6 @@
 // See https://vite-plugin-ssr.com/pageContext-anywhere
 
 import React, { useContext } from 'react'
-// import PropTypes from 'prop-types'
-// import { childrenPropType } from './PropTypeValues'
 import type { PageContext } from './types'
 
 export { PageContextProvider }
@@ -13,11 +11,6 @@ export { usePageContext }
 const Context = React.createContext<PageContext>(
     undefined as unknown as PageContext
 )
-
-// PageContextProvider.propTypes = {
-//     pageContext: PropTypes.any,
-//     children: childrenPropType,
-// }
 
 function PageContextProvider({
     pageContext,
