+++ src/services/updateUserToFirestore.service.jsx
@@ -7,7 +7,7 @@ import { sharingInformationService } from '#@/services/sharing-information'
 export {updateUserToFirestore}
 
 const updateUserToFirestore = (props) => {
-    const user = auth.currentUser || {}
+    const user = auth?.currentUser || {}
     const _firestore = firestore
 
     const { firestoreUserID, userSelectedRol, userEditInfo } = props
